refactor filters, add env scaler to PM

+++ stm32synth/Core/Inc/OSC.h
@@ -15,7 +15,7 @@ typedef enum {
 	LFO_TREMOLO, LFO_PITCH, LFO_FLUTTER
 } OSC_mode_enum;
 
-uint32_t OSC_apply(uint16_t mixer_in, float sample_in, OSC_mode_enum mode);
+float OSC_apply(uint16_t mixer_in, float sample_in, OSC_mode_enum mode);
 
 uint16_t OSC_res_update();
 #endif /* INC_OSC_H_ */
